requirements.txt
Flask==2.2.2
tensorflow
opencv-python
Pillow
requests
scipy
matplotlib
scikit-learn